cell-service 수정 및 호환 체크

+++ lab/capstone-cell-service/src/main/java/com/example/capstonecellservice/dto/CellDto.java
@@ -7,9 +7,9 @@ public class CellDto {
 
     private String cellId;
 
-    private Long totalCell;
-    private Long liveCell;
-    private Long deadCell;
+    private Integer totalCell;
+    private Integer liveCell;
+    private Integer deadCell;
     private Double viability;
 
     private String userId;
