+++ lab/capstone-cell-service/src/main/java/com/example/capstonecellservice/vo/RequestCell.java
@@ -1,13 +1,12 @@
 package com.example.capstonecellservice.vo;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
 @Data
 public class RequestCell {
 
-    private Long totalCell;
-    private Long liveCell;
-    private Long deadCell;
+    private Integer totalCell;
+    private Integer liveCell;
+    private Integer deadCell;
 
 }
