+++ lab/capstone-cell-service/src/main/java/com/example/capstonecellservice/vo/ResponseCell.java
@@ -11,9 +11,9 @@ public class ResponseCell {
 
     private String cellId;
 
-    private Long totalCell;
-    private Long liveCell;
-    private Long deadCell;
+    private Integer totalCell;
+    private Integer liveCell;
+    private Integer deadCell;
     private Double viability;
 
     private Date createAt;
